requirements.txt
# ============================================================
# Autonomous CI/CD Healing Intelligence Core
# requirements.txt
# ============================================================

# Core AI/LLM
openai>=1.30.0
google-generativeai>=0.8.0
langchain>=0.2.0
langchain-openai>=0.1.8
langgraph>=0.1.14
langchain-core>=0.2.0

# AST & Code Analysis
astroid>=3.1.0
pylint>=3.2.0
pyflakes>=3.2.0
black>=24.4.2
isort>=5.13.2

# GitHub Integration
PyGithub>=2.3.0
gitpython>=3.1.43

# Sandbox / Docker
docker>=7.0.0

# Testing Infrastructure
pytest>=8.2.0
pytest-asyncio>=0.23.7
pytest-cov>=5.0.0
pytest-json-report>=1.5.0
coverage>=7.5.0

# Data & Validation
pydantic>=2.7.0
pydantic-settings>=2.2.1

# Async / Concurrency
asyncio-throttle>=1.0.2
anyio>=4.4.0

# Utilities
python-dotenv>=1.0.1
rich>=13.7.1
loguru>=0.7.2
tenacity>=8.3.0
tqdm>=4.66.4
networkx>=3.3
toml>=0.10.2
packaging>=24.0

# JSON / YAML
pyyaml>=6.0.1
jsonschema>=4.22.0

# API Layer
fastapi>=0.110.0
uvicorn>=0.29.0
requests>=2.31.0
